refactor: Outsource into getOrCreateNewMediaFile method
